+++ src/ecs/systems/systems.ts
@@ -3,20 +3,24 @@ import { _k } from "../../shared";
 export type System = {
     name: string;
     run: () => void;
-    when: LCEvents[];
+    when: SystemPhase[];
 };
 
 // Lifecycle events
-export enum LCEvents {
-    "BeforeUpdate",
-    "BeforeFixedUpdate",
-    "BeforeDraw",
-    "AfterUpdate",
-    "AfterFixedUpdate",
-    "AfterDraw",
+export enum SystemPhase {
+    BeforeUpdate,
+    BeforeFixedUpdate,
+    BeforeDraw,
+    AfterUpdate,
+    AfterFixedUpdate,
+    AfterDraw,
 }
 
-export const system = (name: string, action: () => void, when: LCEvents[]) => {
+export const system = (
+    name: string,
+    action: () => void,
+    when: SystemPhase[],
+) => {
     const systems = _k.game.systems;
     const replacingSystemIdx = systems.findIndex((s) => s.name === name);